Update README.md
Project Deliverables
You will be required to complete the following deliverable.
● A python notebook with your solution.
Instructions
Background Information
Together with a team of startup entrepreneurs, you decide to work on an idea that could
change the way people search for jobs. You decide that job scraping could be the next
big thing as there are actively many people looking for jobs in the country, in this case,
Kenya.
Problem Statement
The problem is that there are many job listings which can not get visits for the target job
seekers. While working in a team, your task as a data scientist for this project is to
scrape for job titles and links and then put them in a single table that can be used by
your team members to further build a job aggregator.
You will be required to scrape for data from the following three technology webpages:
● PigiaMe: https://www.pigiame.co.ke/it-software-jobs
● MyJobMag: https://www.myjobmag.co.ke/jobs-by-field/information-technology
● KenyaJob:
https://www.kenyajob.com/job-vacancies-search-kenya?f%5B0%5D=im_field_offr
e_secteur%3A133
Your deliverable will be a python script that performs the mentioned task. You can use
the following guiding notebook to get started [Link].